+++ libraries/SysAPI/include/sys/Pipe.hpp
@@ -16,11 +16,9 @@ public:
   Pipe &operator=(Pipe &&) = default;
 
   fs::File &read_file() { return m_read_file; }
-
   const fs::File &read_file() const { return m_read_file; }
 
   fs::File &write_file() { return m_write_file; }
-
   const fs::File &write_file() const { return m_write_file; }
 
 private: