+++ CHANGES.md
@@ -3,6 +3,7 @@
 
 ## New Features
 
+- `sys::Process` no longer provides a pipe. It gives an `fs::DataFile` for standard output and error.
 - Add `sys::System::launch_browser()` to open a link in the system browser
 - Add `api::catch_segmentation_fault()` to call `API_ASSERT()` when a segmentation fault happens (desktop)